prevent members query when mentions disabled
