+++ src/components/MessageInput/DefaultTriggerProvider.tsx
@@ -1,13 +1,22 @@
-import React from 'react';
-import {
-  MessageInputContextProvider,
-  useMessageInputContext,
-} from '../../context/MessageInputContext';
+import React, { PropsWithChildren } from 'react';
+
 import { useCommandTrigger } from './hooks/useCommandTrigger';
 import { useEmojiTrigger } from './hooks/useEmojiTrigger';
 import { useUserTrigger } from './hooks/useUserTrigger';
 
-import type { TriggerSettings } from '../ChatAutoComplete/ChatAutoComplete';
+import {
+  MessageInputContextProvider,
+  useMessageInputContext,
+} from '../../context/MessageInputContext';
+
+import type { EmojiData } from 'emoji-mart';
+import type { UserFilters, UserOptions, UserSort } from 'stream-chat';
+
+import type { SuggestionCommand, SuggestionUser } from '../ChatAutoComplete/ChatAutoComplete';
+import type { CommandItemProps } from '../CommandItem/CommandItem';
+import type { EmoticonItemProps } from '../EmoticonItem/EmoticonItem';
+import type { UserItemProps } from '../UserItem/UserItem';
+
 import type {
   CustomTrigger,
   DefaultAttachmentType,
@@ -17,7 +26,64 @@ import type {
   DefaultMessageType,
   DefaultReactionType,
   DefaultUserType,
+  UnknownType,
 } from '../../types/types';
+
+export type AutocompleteMinimalData = {
+  id?: string;
+  name?: string;
+} & ({ id: string } | { name: string });
+
+export type CommandTriggerSetting<
+  Co extends DefaultCommandType = DefaultCommandType
+> = TriggerSetting<CommandItemProps, SuggestionCommand<Co>>;
+
+export type EmojiTriggerSetting = TriggerSetting<EmoticonItemProps, EmojiData>;
+
+export type UserTriggerSetting<Us extends DefaultUserType<Us> = DefaultUserType> = TriggerSetting<
+  UserItemProps,
+  SuggestionUser<Us>
+>;
+
+export type MentionQueryParams<Us extends DefaultUserType<Us> = DefaultUserType> = {
+  filters?: UserFilters<Us>;
+  options?: UserOptions;
+  sort?: UserSort<Us>;
+};
+
+export type TriggerSetting<T extends UnknownType = UnknownType, U = UnknownType> = {
+  component: string | React.ComponentType<T>;
+  dataProvider: (
+    query: string,
+    text: string,
+    onReady: (data: (U & AutocompleteMinimalData)[], token: string) => void,
+  ) => U[] | Promise<void> | void;
+  output: (
+    entity: U,
+  ) =>
+    | {
+        caretPosition: 'start' | 'end' | 'next' | number;
+        text: string;
+        key?: string;
+      }
+    | string
+    | null;
+  callback?: (item: U) => void;
+};
+
+export type TriggerSettings<
+  Co extends DefaultCommandType = DefaultCommandType,
+  Us extends DefaultUserType<Us> = DefaultUserType,
+  V extends CustomTrigger = CustomTrigger
+> =
+  | {
+      [key in keyof V]: TriggerSetting<V[key]['componentProps'], V[key]['data']>;
+    }
+  | {
+      '/': CommandTriggerSetting<Co>;
+      ':': EmojiTriggerSetting;
+      '@': UserTriggerSetting<Us>;
+    };
 
 export const DefaultTriggerProvider = <
   At extends DefaultAttachmentType = DefaultAttachmentType,
@@ -30,17 +96,18 @@ export const DefaultTriggerProvider = <
   V extends CustomTrigger = CustomTrigger
 >({
   children,
-}: React.PropsWithChildren<Record<string, unknown>>) => {
+}: PropsWithChildren<Record<string, unknown>>) => {
   const currentValue = useMessageInputContext<At, Ch, Co, Ev, Me, Re, Us, V>();
 
   const defaultAutocompleteTriggers: TriggerSettings<Co, Us> = {
     '/': useCommandTrigger<At, Ch, Co>(),
     ':': useEmojiTrigger(currentValue.emojiIndex),
-    '@': useUserTrigger<At, Ch, Co, Ev, Me, Re, Us>(
-      currentValue.mentionQueryParams,
-      currentValue.onSelectUser,
-      currentValue.mentionAllAppUsers,
-    ),
+    '@': useUserTrigger<At, Ch, Co, Ev, Me, Re, Us>({
+      disableMentions: currentValue.disableMentions,
+      mentionAllAppUsers: currentValue.mentionAllAppUsers,
+      mentionQueryParams: currentValue.mentionQueryParams,
+      onSelectUser: currentValue.onSelectUser,
+    }),
   };
 
   const newValue = {
