+++ src/version.ts
@@ -1,1 +1,1 @@
-export const version = '6.0.3';
+export const version = '6.0.4';
